The architecture of the code has been remade + The jacoco coverage has been made
