+++ TourGuide/src/main/java/tourGuide/DTO/UserDTO.java
@@ -1,4 +1,4 @@
-package tourGuide.user;
+package tourGuide.DTO;
 
 import gpsUtil.location.Location;
 import gpsUtil.location.VisitedLocation;
