+++ TourGuide/src/main/java/tourGuide/model/user/UserReward.java
@@ -1,0 +1,30 @@
+package tourGuide.model.user;
+
+import gpsUtil.location.Attraction;
+import gpsUtil.location.VisitedLocation;
+
+public class UserReward {
+
+	public final VisitedLocation visitedLocation;
+	public final Attraction attraction;
+	private int rewardPoints;
+	public UserReward(VisitedLocation visitedLocation, Attraction attraction, int rewardPoints) {
+		this.visitedLocation = visitedLocation;
+		this.attraction = attraction;
+		this.rewardPoints = rewardPoints;
+	}
+	
+	public UserReward(VisitedLocation visitedLocation, Attraction attraction) {
+		this.visitedLocation = visitedLocation;
+		this.attraction = attraction;
+	}
+
+	public void setRewardPoints(int rewardPoints) {
+		this.rewardPoints = rewardPoints;
+	}
+	
+	public int getRewardPoints() {
+		return rewardPoints;
+	}
+	
+}
